refactored with a new package
